Nathan db assignment 3

+++ code/db/db-assignment-2/routes/Recipe.js
@@ -1,15 +1,15 @@
 import { Router } from "express";
 import Recipe from "../models/recipeModel.js";
+import authMiddleware from "../middlewares/authMiddleware.js";
 
 const route = Router();
 
 route.get("/", async (req, res) => {
   try {
     const recipes = await Recipe.find();
-    console.log(recipes);
     res.json(recipes);
   } catch (error) {
-    console.log("Error fetching all recipes");
+    console.log(error);
   }
 });
 
@@ -23,13 +23,15 @@ route.get("/:id", async (req, res) => {
   }
 });
 
-route.post("/", async (req, res) => {
+route.post("/", authMiddleware, async (req, res) => {
+  const { userId } = req.user;
   try {
     const newRecipe = {
       title: req.body.title,
       author: req.body.author,
       ingredients: req.body.ingredients,
       instructions: req.body.instructions,
+      createdBy: userId,
     };
     const recipe = await Recipe.create(newRecipe);
     res.status(201).json(recipe);
@@ -38,9 +40,21 @@ route.post("/", async (req, res) => {
   }
 });
 
-route.put("/:id", async (req, res) => {
+route.put("/:id", authMiddleware, async (req, res) => {
+  const { userId } = req.user;
+  const { id } = req.params;
+
   try {
-    const { id } = req.params;
+    const FoundRecipe = await Recipe.findById(id);
+    const createdbyId = FoundRecipe.createdBy.toString();
+    console.log("Found Recipe", typeof createdbyId);
+    if (userId !== createdbyId) {
+      return res.status(400).json({
+        message:
+          "Unauthorized Access. User did not create this recipe and therefore cannout update it",
+      });
+    }
+
     const updatedRecipe = {
       title: req.body.title,
       author: req.body.author,
@@ -54,15 +68,31 @@ route.put("/:id", async (req, res) => {
   }
 });
 
-route.delete("/:id", async (req, res) => {
+route.delete("/:id", authMiddleware, async (req, res) => {
+  const { userId } = req.user;
+  const { id } = req.params;
+
   try {
-    const { id } = req.params;
-    
+    const FoundRecipe = await Recipe.findById(id);
+
+    if (!FoundRecipe) {
+      return res.status(400).json({ message: "No Recipe with that Id found" });
+    }
+
+    const createdbyId = FoundRecipe.createdBy.toString();
+
+    if (userId !== createdbyId) {
+      return res.status(400).json({
+        message:
+          "Unauthorized Access. User did not create this recipe and therefore cannout delete it",
+      });
+    }
+
     const deletedrecipe = await Recipe.findByIdAndDelete(id);
-    console.log("Deleted Recipe", deletedrecipe)
+    console.log("Deleted Recipe", deletedrecipe);
     res.status(200).json(deletedrecipe);
   } catch (error) {
-    console.log("Error deleting recipe");
+    console.log("Error deleting recipe", error);
   }
 });
